Consolidate OSDU test suite: 16 files → 7
RDDMS tests (test_rddms.py):
- Merged double_roundtrip + discovery + local ETP roundtrip into one file
- Deleted redundant: test_etp_eqndev_roundtrip, test_etp_epc_crossformat,
  test_eqndev_roundtrip (all subsumed by double roundtrip)
- All RDDMS tests excluded from CI via requires_rddms marker

EPC compliance (test_epc_compliance.py):
- Merged grid3d_compliance + surface_xyz_compliance + well_trisurface
- Deleted test_epc_roundtrip.py (all tests were strict subsets)

Other merges:
- test_metadata_roundtrip.py → merged into test_osdu_unit.py
- test_resqpy_interop.py + test_resqpy_pipeline.py → test_resqpy.py
- test_epc_grid3d_operations.py → renamed test_epc_operations.py

Result: 193 CI tests pass (was 205, removed 12 redundant), 0 failures
